+++ src/OpcUaStackCore/SecureChannel/SecureChannelBase.cpp
@@ -34,8 +34,6 @@ namespace OpcUaStackCore
 	// ------------------------------------------------------------------------
 	SecureChannelBase::SecureChannelBase(SecureChannelType secureChannelType)
 	: secureChannelType_(secureChannelType)
-	, cryptoManager_()
-	, applicationCertificate_()
 	{
 	}
 
@@ -689,8 +687,8 @@ namespace OpcUaStackCore
 	void
 	SecureChannelBase::asyncWriteOpenSecureChannelResponse(SecureChannel* secureChannel)
 	{
-		assert(applicationCertificate_.get() != nullptr);
-		assert(applicationCertificate_->certificate().get() != nullptr);
+		assert(applicationCertificate().get() != nullptr);
+		assert(applicationCertificate()->certificate().get() != nullptr);
 
 		SecureChannelSecuritySettings& securitySettings = secureChannel->securitySettings();
 
@@ -726,7 +724,7 @@ namespace OpcUaStackCore
 		securityHeader.senderCertificate().reset();
 		if (securityHeader.isSignatureEnabled()) {
 			// FIXME: use sender certificate chain
-			applicationCertificate_->certificate()->toDERBuf(securityHeader.senderCertificate());
+			applicationCertificate()->certificate()->toDERBuf(securityHeader.senderCertificate());
 		}
 		securityHeader.receiverCertificateThumbprint().reset();
 		if (securityHeader.isEncryptionEnabled()) {
@@ -1602,157 +1600,6 @@ namespace OpcUaStackCore
 	//
 	// ------------------------------------------------------------------------
 	// ------------------------------------------------------------------------
-	void
-	SecureChannelBase::cryptoManager(CryptoManager::SPtr& cryptoManager)
-	{
-		cryptoManager_ = cryptoManager;
-	}
-
-	void
-	SecureChannelBase::applicationCertificate(ApplicationCertificate::SPtr& applicationCertificate)
-	{
-		applicationCertificate_ = applicationCertificate;
-	}
-
-	OpcUaStatusCode
-	SecureChannelBase::secureReceivedOpenSecureChannelRequest(
-		SecureChannel* secureChannel
-	)
-	{
-		OpcUaStatusCode statusCode;
-
-		SecureChannelSecuritySettings& securitySettings = secureChannel->securitySettings();
-		SecurityHeader* securityHeader = &secureChannel->securityHeader_;
-
-		// check if encryption or signature is enabled
-		if (!securityHeader->isEncryptionEnabled() && !securityHeader->isSignatureEnabled()) {
-			return Success;
-		}
-
-		// find crypto base
-		CryptoBase::SPtr cryptoBase = cryptoManager_->get(securityHeader->securityPolicyUri().toString());
-		if (cryptoBase.get() == nullptr) {
-			Log(Error, "crypto base not available for security policy uri")
-				.parameter("SecurityPolicyUri", securityHeader->securityPolicyUri().toString());
-			return BadSecurityPolicyRejected;
-		}
-		cryptoBase->isLogging(secureChannel->isLogging_);
-		securitySettings.cryptoBase(cryptoBase);
-
-		// decrypt received open secure channel request
-		if (securityHeader->isEncryptionEnabled()) {
-			statusCode = decryptReceivedOpenSecureChannel(secureChannel);
-			if (statusCode != Success) {
-				return statusCode;
-			}
-		}
-
-		// verify signature
-		if (securityHeader->isSignatureEnabled()) {
-			Certificate::SPtr partnerCertificate = securityHeader->certificateChain().getCertificate();
-			securitySettings.partnerCertificate(partnerCertificate);
-			statusCode = verifyReceivedOpenSecureChannel(secureChannel);
-			if (statusCode != Success) {
-				return statusCode;
-			}
-		}
-
-		return Success;
-	}
-
-	OpcUaStatusCode
-	SecureChannelBase::decryptReceivedOpenSecureChannel(
-		SecureChannel* secureChannel
-	)
-	{
-		SecureChannelSecuritySettings& securitySettings = secureChannel->securitySettings();
-		SecurityHeader* securityHeader = &secureChannel->securityHeader_;
-
-		uint32_t receivedDataLen = secureChannel->recvBuffer_.size();
-		OpcUaStatusCode statusCode;
-
-		// check receiver certificate
-		if (securityHeader->receiverCertificateThumbprint() != applicationCertificate_->certificate()->thumbPrint()) {
-			Log(Error, "receiver certificate invalid")
-				.parameter("ReceiverCertificateThumbprint", securityHeader->receiverCertificateThumbprint());
-			return BadCertificateInvalid;
-		}
-
-		// the number of received bytes must be a multiple of the key length
-		if (receivedDataLen % (applicationCertificate_->privateKey()->keySize()/8) != 0) {
-			Log(Error, "number of received bytes invalid")
-				.parameter("ReceivedDataLen", receivedDataLen);
-			return BadSecurityChecksFailed;
-		}
-
-		// decrypt received buffer
-		std::iostream ios(&secureChannel->recvBuffer_);
-		MemoryBuffer encryptedText(receivedDataLen);
-		MemoryBuffer plainText(receivedDataLen);
-		ios.read(encryptedText.memBuf(), receivedDataLen);
-
-		statusCode = securitySettings.cryptoBase()->asymmetricDecrypt(
-			encryptedText.memBuf(),
-			encryptedText.memLen(),
-			*applicationCertificate_->privateKey().get(),
-			plainText.memBuf(),
-			&receivedDataLen
-		);
-		if (statusCode != Success) {
-			Log(Error, "decrypt open secure channel request error")
-				.parameter("StatusCode", OpcUaStatusCodeMap::shortString(statusCode));
-			return BadSecurityChecksFailed;
-		}
-
-		ios.write(plainText.memBuf(), receivedDataLen);
-		return Success;
-	}
-
-	OpcUaStatusCode
-	SecureChannelBase::verifyReceivedOpenSecureChannel(
-		SecureChannel* secureChannel
-	)
-	{
-		OpcUaStatusCode statusCode;
-
-		SecureChannelSecuritySettings& securitySettings = secureChannel->securitySettings();
-		MessageHeader* messageHeader = &secureChannel->messageHeader_;
-		SecurityHeader* securityHeader = &secureChannel->securityHeader_;
-
-		// get public key client certificate
-		PublicKey publicKey = securityHeader->certificateChain().getCertificate()->publicKey();
-		uint32_t signTextLen = publicKey.keySizeInBytes();
-
-		// create plain text buffer (with signature at end of buffer)
-		boost::asio::streambuf streambuf;
-		std::iostream os(&streambuf);
-		messageHeader->opcUaBinaryEncode(os, true);
-		securityHeader->opcUaBinaryEncode(os);
-
-		uint32_t plainTextLen = streambuf.size() + secureChannel->recvBuffer_.size();
-		MemoryBuffer plainText(plainTextLen);
-
-		const char* header = boost::asio::buffer_cast<const char*>(streambuf.data());
-		memcpy(plainText.memBuf(), header, streambuf.size());
-		const char* body = boost::asio::buffer_cast<const char*>(secureChannel->recvBuffer_.data());
-		memcpy(plainText.memBuf()+streambuf.size(), body, secureChannel->recvBuffer_.size());
-
-		// verify signature
-		statusCode = securitySettings.cryptoBase()->asymmetricVerify(
-			plainText.memBuf(),
-			plainText.memLen() - signTextLen,
-			publicKey,
-			plainText.memBuf() + plainText.memLen() - signTextLen,
-			signTextLen
-		);
-		if (statusCode != Success) {
-			Log(Error, "verify open secure channel request error")
-				.parameter("StatusCode", OpcUaStatusCodeMap::shortString(statusCode));
-			return BadSecurityChecksFailed;
-		}
-
-		return Success;
-	}
 
 	OpcUaStatusCode
 	SecureChannelBase::secureSendOpenSecureChannelResponse(
@@ -1801,8 +1648,8 @@ namespace OpcUaStackCore
 		OpcUaStatusCode statusCode;
 
 		SecureChannelSecuritySettings& securitySettings = secureChannel->securitySettings();
-		PublicKey publicKey = applicationCertificate_->certificate()->publicKey();
-		PrivateKey::SPtr privateKey = applicationCertificate_->privateKey();
+		PublicKey publicKey = applicationCertificate()->certificate()->publicKey();
+		PrivateKey::SPtr privateKey = applicationCertificate()->privateKey();
 
 		// get asymmetric key length
 		uint32_t asymmetricKeyLen = 0;
