+++ src/OpcUaStackCore/SecureChannel/SecureChannelBase.h
@@ -20,7 +20,7 @@
 #define __OpUaStackCore_SecureChannelBase_h__
 
 #include "OpcUaStackCore/Base/os.h"
-#include "OpcUaStackCore/SecureChannel/SecureChannel.h"
+#include "OpcUaStackCore/SecureChannel/SecureChannelCrypto.h"
 #include "OpcUaStackCore/SecureChannel/HelloMessage.h"
 #include "OpcUaStackCore/SecureChannel/AcknowledgeMessage.h"
 #include "OpcUaStackCore/SecureChannel/SecurityHeader.h"
@@ -30,13 +30,12 @@
 #include "OpcUaStackCore/SecureChannel/OpenSecureChannelResponse.h"
 #include "OpcUaStackCore/SecureChannel/CloseSecureChannelRequest.h"
 #include "OpcUaStackCore/SecureChannel/CloseSecureChannelResponse.h"
-#include "OpcUaStackCore/Certificate/CryptoManager.h"
-#include "OpcUaStackCore/Certificate/ApplicationCertificate.h"
 
 namespace OpcUaStackCore
 {
 
 	class DLLEXPORT SecureChannelBase
+	: public SecureChannelCrypto
 	{
 	  public:
 		typedef enum
@@ -115,20 +114,11 @@ namespace OpcUaStackCore
 
 		void asyncRead(SecureChannel* secureChannel);
 
-		void cryptoManager(CryptoManager::SPtr& cryptoManager);
-		void applicationCertificate(ApplicationCertificate::SPtr& applicationCertificate);
-
 	  private:
-		OpcUaStatusCode secureReceivedOpenSecureChannelRequest(
-			SecureChannel* secureChannel
-		);
-		OpcUaStatusCode decryptReceivedOpenSecureChannel(
-			SecureChannel* secureChannel
-		);
-		OpcUaStatusCode verifyReceivedOpenSecureChannel(
-			SecureChannel* secureChannel
-		);
 
+		//
+		// open secure channel response
+		//
 		OpcUaStatusCode secureSendOpenSecureChannelResponse(
 			MemoryBuffer& plainText,
 			MemoryBuffer& encryptedText,
@@ -144,6 +134,9 @@ namespace OpcUaStackCore
 			SecureChannel* secureChannel
 		);
 
+		//
+		// message request
+		//
 		OpcUaStatusCode secureReceivedMessageRequest(
 			SecureChannel* secureChannel
 		);
@@ -154,6 +147,9 @@ namespace OpcUaStackCore
 			SecureChannel* secureChannel
 		);
 
+		//
+		// message response
+		//
 		OpcUaStatusCode secureSendMessageResponse(
 			MemoryBuffer& plainText,
 			MemoryBuffer& encryptedText,
@@ -217,9 +213,6 @@ namespace OpcUaStackCore
 		);
 
 		SecureChannelType secureChannelType_;
-
-		CryptoManager::SPtr cryptoManager_;
-		ApplicationCertificate::SPtr applicationCertificate_;
 	};
 
 }
